+++ RELEASES.md
@@ -4,7 +4,7 @@
 
 #### New features
 
-- 
+- Added Generalized Wasserstein Barycenter solver + example (PR #372)
 
 #### Closed issues